Seperate JSON Parser Function to File Reader
